+++ app/interface/src/pages/Builder/index.tsx
@@ -1,6 +1,7 @@
 import Terminal from "../../components/Terminal";
 import Tfvar from "../../components/Tfvar";
 import { useServerStatus } from "../../hooks/useServerStatus";
+import PageLayout from "../../layouts/PageLayout";
 import ServerError from "../ServerError";
 
 export default function Builder() {
@@ -11,12 +12,9 @@ export default function Builder() {
   }
 
   return (
-    <div>
-      <p className="mb-6 border-b-2 border-slate-500 py-4 text-4xl">
-        Lab Builder
-      </p>
+    <PageLayout heading="Lab Builder">
       <Tfvar />
       <Terminal />
-    </div>
+    </PageLayout>
   );
 }
