+++ app/interface/src/pages/Landing/index.tsx
@@ -1,6 +1,7 @@
 import { useState } from "react";
 import { MdOutlineContentCopy, MdDoneOutline } from "react-icons/md";
 import { Link } from "react-router-dom";
+import PageLayout from "../../layouts/PageLayout";
 
 export default function Landing() {
   const [copy, setCopy] = useState<boolean>(false);
@@ -15,7 +16,7 @@ export default function Landing() {
     }, 3000);
   }
   return (
-    <div className="gap-y-10">
+    <PageLayout>
       <h1 className="py-10 text-8xl font-bold">Head start your lab repros.</h1>
       <p className="pb-20 text-3xl text-slate-500">
         All you need to do is install docker 🐋, run server locally on your
@@ -47,6 +48,6 @@ export default function Landing() {
           </button>
         </div>
       </div>
-    </div>
+    </PageLayout>
   );
 }
